change noise results to doubles, update example noise usage

+++ src/perlin_noise.cpp
@@ -5,7 +5,7 @@
 #include <iomanip>
 
 
-float PerlinNoise::interpolate(float a, float b, float weight) {
+double PerlinNoise::interpolate(double a, double b, double weight) {
     if (weight < 0) return a;
     if (weight > 1) return b;
     return (b - a) * ((weight * (weight * 6.0 - 15.0) + 10.0) * weight * weight * weight) + a;
@@ -20,27 +20,27 @@ Vector2 PerlinNoise::randomGradient(int x, int y) {
     return v;
 };
 
-float PerlinNoise::dotProduct(int xGrid, int yGrid, float px, float py) {
+double PerlinNoise::dotProduct(int xGrid, int yGrid, double px, double py) {
     Vector2 gradient = randomGradient(xGrid, yGrid);
     // get the offset vector from the grid point to the target point
     Vector2 offsetVector = {
-        px - (float)xGrid,      // x
-        py - (float)yGrid       // y
+        px - (double)xGrid,      // x
+        py - (double)yGrid       // y
     };
     return gradient.x * offsetVector.x + gradient.y * offsetVector.y;
 };
 
-float PerlinNoise::noise(float x, float y) {
+double PerlinNoise::noise(double x, double y) {
     int xGrid0 = (int)floor(x);
     int yGrid0 = (int)floor(y);
     int xGrid1 = xGrid0 + 1;
     int yGrid1 = yGrid0 + 1;
 
     // calculate weights
-    float wx = x - (float)xGrid0;
-    float wy = y - (float)yGrid0;
+    double wx = x - (double)xGrid0;
+    double wy = y - (double)yGrid0;
 
-    float dot1, dot2, interp1, interp2;
+    double dot1, dot2, interp1, interp2;
     dot1 = dotProduct(xGrid0, yGrid0, x, y);
     dot2 = dotProduct(xGrid1, yGrid0, x, y);
     interp1 = interpolate(dot1, dot2, wx);
@@ -62,12 +62,14 @@ float PerlinNoise::noise(float x, float y) {
 //     std::cout << n2 << "\n";
 //     std::cout << n3 << "\n";
 
-//     for (float i = 44; i < 45; i += 0.1) {
-//         for (float j = 44; j < 45; j += 0.1) {
-//             float val = PerlinNoise::noise(i, j);
+//     for (double i = 0.05; i < 1; i += 0.1) {
+//         for (double j = 0.05; j < 1; j += 0.1) {
+//             double val = PerlinNoise::noise(i, j);
+//             val = (val + 1)/2;
+//             val = (int)(val * 100);
 //             std::cout << std::left << std::setw(12) << val << " ";
+//             // std::cout << std::left << std::setw(12) << val << " ";
 //         }
 //         std::cout << "\n";
 //     }
-
 // }
